Revert "Some more errors from online Doctests cleared"
This reverts commit 26b24f2d6f8ef6817504d3cf38fd825a2a210736.

# Conflicts:
#	pandas/core/indexes/base.py

+++ pandas/core/arrays/sparse/array.py
@@ -359,7 +359,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
     [0, 0, 1, 2]
     Fill: 0
     IntIndex
-    Indices: array([2, 3], dtype=int32)
+    Indices: array([2, 3])
     """
 
     _subtyp = "sparse_array"  # register ABCSparseArray
@@ -642,7 +642,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         --------
         >>> s = SparseArray([0, 0, 1, 0, 2], fill_value=0)
         >>> s.sp_values
-        array([1, 2])
+        array([1, 2], dtype=int64)
         """
         return self._sparse_values
 
@@ -1280,7 +1280,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         [0, 0, 1, 2]
         Fill: 0
         IntIndex
-        Indices: array([2, 3], dtype=int32)
+        Indices: array([2, 3])
 
         >>> arr.astype(SparseDtype(np.dtype('int32')))
         [0, 0, 1, 2]
@@ -1358,7 +1358,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         [10, 11, 12]
         Fill: 10
         IntIndex
-        Indices: array([1, 2], dtype=int32)
+        Indices: array([1, 2])
 
         >>> arr.map({0: 10, 1: 11, 2: 12})
         [10, 11, 12]
