+++ pandas/core/ops/missing.py
@@ -95,7 +95,7 @@ def mask_zero_div_zero(x, y, result: np.ndarray) -> np.ndarray:
     --------
     >>> x = np.array([1, 0, -1], dtype=np.int64)
     >>> x
-    array([ 1,  0, -1])
+    array([ 1,  0, -1], dtype=int64)
     >>> y = 0       # int 0; numpy behavior is different with float
     >>> result = x // y
     >>> result      # raw numpy result does not fill division by zero
